change prime-erica
change file name and gap
